update:重构部分分层逻辑
