+++ src/apps/api/user/delete.go
@@ -1,7 +1,7 @@
 package user
 
 import (
-	"alemongo/src/apps/token"
+	"alemongo/middlewares"
 	"alemongo/src/users"
 	"net/http"
 
@@ -10,7 +10,7 @@ import (
 
 // method: DELETE
 func Delete(ctx *gin.Context) {
-	adminname, exists := token.GetUserName(ctx)
+	adminname, exists := middlewares.GetUserName(ctx)
 	if !exists {
 		ctx.JSON(http.StatusBadRequest, gin.H{
 			"code": http.StatusBadRequest,
